+++ WorldEditCommands/Object/FieldAutoComplete.cs
@@ -9,16 +9,44 @@ using UnityEngine;
 
 namespace WorldEditCommands;
 
+class ObjectHash { }
+class LocationHash { }
+class RoomHash { }
+
 public class FieldAutoComplete
 {
   public static Dictionary<string, Dictionary<string, Type>> Fields = [];
+  public static Dictionary<string, List<string>> Keys = [];
   private static List<string> Components = [];
 
   public static void Init()
   {
     // Easier to always load everything at start up, doesn't take long anyways.
     Fields = LoadFields();
-    Components = [.. Fields.Keys];
+    InitZdoFields();
+    foreach (var kvp in Fields)
+    {
+      if (!Keys.ContainsKey(kvp.Key))
+        Keys[kvp.Key] = [];
+      foreach (var kvp2 in kvp.Value)
+      {
+        var key = kvp2.Key.Replace("m_", "");
+        if (Keys[kvp.Key].Contains(key)) continue;
+        Keys[kvp.Key].Add(key);
+      }
+    }
+    foreach (var kvp in ZdoFields)
+    {
+      if (!Keys.ContainsKey(kvp.Key))
+        Keys[kvp.Key] = [];
+      foreach (var kvp2 in kvp.Value)
+      {
+        if (Keys[kvp.Key].Contains(kvp2.Key)) continue;
+        Keys[kvp.Key].Add(kvp2.Key);
+      }
+    }
+    Components = [.. Keys.Keys];
+    Components = Components.Distinct().ToList();
   }
   private static readonly HashSet<Type> ValidTypes = [
     typeof(string),
@@ -105,7 +133,7 @@ public class FieldAutoComplete
   private static List<string> GetComponents(string prefab)
   {
     if (ZNetScene.instance.m_namedPrefabs.TryGetValue(prefab.GetStableHashCode(), out var gameObject))
-      return gameObject.GetComponentsInChildren<MonoBehaviour>().Select(c => c.GetType().Name).ToList();
+      return [.. gameObject.GetComponentsInChildren<MonoBehaviour>().Select(c => c.GetType().Name), "zdo"];
     return Components;
   }
   public static List<string> GetFields()
@@ -125,7 +153,7 @@ public class FieldAutoComplete
   }
   private static List<string> GetFields(string component)
   {
-    return Fields.TryGetValue(component, out var fields) ? [.. fields.Keys.Select(s => s.Replace("m_", ""))] : [];
+    return Keys.TryGetValue(component, out var keys) ? keys : [];
   }
   public static List<string> GetTypes(int index)
   {
@@ -133,15 +161,19 @@ public class FieldAutoComplete
     var command = GetInput();
     var prefab = PrefabFromCommand(command);
     var component = ComponentFromCommand(prefab, command);
-    if (!Fields.TryGetValue(component, out var fields)) return [];
     var field = FieldFromCommand(component, command);
-    if (!fields.TryGetValue(field, out var type)) return [];
+    var realField = RealField(component, field, out var zdoField);
+    var type = GetType(component, realField, zdoField);
     if (type == typeof(string)) return ["Text"];
     if (type == typeof(int)) return ["Number"];
     if (type == typeof(float)) return ["Decimal"];
     if (type == typeof(bool)) return ["true", "false"];
     if (type == typeof(Vector3)) return ServerDevcommands.ParameterInfo.XZY("Field", index);
+    if (type == typeof(Quaternion)) return ServerDevcommands.ParameterInfo.YXZ("Field", index);
     if (type == typeof(GameObject)) return tweaks ? GetIdsOrTransforms(prefab, component, field) : ServerDevcommands.ParameterInfo.Ids;
+    if (type == typeof(ObjectHash)) return ServerDevcommands.ParameterInfo.ObjectIds;
+    if (type == typeof(LocationHash)) return ServerDevcommands.ParameterInfo.LocationIds;
+    if (type == typeof(RoomHash)) return ServerDevcommands.ParameterInfo.RoomIds;
     if (tweaks && type == typeof(ItemDrop)) return ServerDevcommands.ParameterInfo.ItemIds;
     if (tweaks && type.IsEnum) return [.. Enum.GetNames(type)];
     if (tweaks && type == typeof(EffectList)) return ServerDevcommands.ParameterInfo.Ids;
@@ -153,7 +185,7 @@ public class FieldAutoComplete
     if (!obj) return ServerDevcommands.ParameterInfo.Ids;
     var c = obj.GetComponents<MonoBehaviour>().FirstOrDefault(c => c.GetType().Name == component);
     if (c == null) return ServerDevcommands.ParameterInfo.Ids;
-    var f = c.GetType().GetField(RealField(component, field));
+    var f = c.GetType().GetField(RealField(component, field, out _));
     if (f == null || f.FieldType != typeof(GameObject)) return ServerDevcommands.ParameterInfo.Ids;
     var go = (GameObject)f.GetValue(c);
     if (go && ZNetScene.instance.GetPrefab(go.name) == go) return ServerDevcommands.ParameterInfo.Ids;
@@ -175,13 +207,22 @@ public class FieldAutoComplete
     var arg = split.LastOrDefault(s => s.StartsWith("field=", StringComparison.Ordinal) || s.StartsWith("f=", StringComparison.Ordinal));
     if (arg == null) return "";
     split = arg.Split('=')[1].Split(',');
-    return split.Length < 2 ? "" : RealField(component, split[1]);
+    return split.Length < 2 ? "" : RealField(component, split[1], out _);
   }
-  public static Type GetType(string component, string field)
+  public static Type GetType(string component, string field, bool zdoField)
   {
-    if (!Fields.TryGetValue(component, out var fields)) return typeof(void);
-    if (!fields.TryGetValue(field, out var type)) return typeof(void);
-    return type;
+    if (zdoField)
+    {
+      if (!ZdoFields.TryGetValue(component, out var zdo)) return typeof(void);
+      if (!zdo.TryGetValue(field, out var type)) return typeof(void);
+      return type;
+    }
+    else
+    {
+      if (!Fields.TryGetValue(component, out var fields)) return typeof(void);
+      if (!fields.TryGetValue(field, out var type)) return typeof(void);
+      return type;
+    }
   }
 
   public static string RealComponent(string prefab, string component)
@@ -191,16 +232,492 @@ public class FieldAutoComplete
     ?? components.FirstOrDefault(s => s.StartsWith(component, StringComparison.OrdinalIgnoreCase))
     ?? component;
   }
-  public static string RealField(string component, string field)
+  public static string RealField(string component, string field, out bool zdoField)
   {
+    zdoField = false;
     var f = $"m_{field}";
     var fields = Fields.TryGetValue(component, out var fs) ? fs.Keys.ToList() : [];
     var primaryField = fields.FirstOrDefault(s => s.Equals(f, StringComparison.OrdinalIgnoreCase));
     if (primaryField != null) return primaryField;
-    var secondaryField = fields.FirstOrDefault(s => s.StartsWith(f, StringComparison.OrdinalIgnoreCase) || s.StartsWith(field, StringComparison.OrdinalIgnoreCase));
+    zdoField = true;
+    var zdoFields = ZdoFields.TryGetValue(component, out var zdo) ? zdo.Keys.ToList() : [];
+    var secondaryField = zdoFields.FirstOrDefault(s => s.Equals(field, StringComparison.OrdinalIgnoreCase));
     if (secondaryField != null) return secondaryField;
+    zdoField = false;
+    var tertiaryField = fields.FirstOrDefault(s => s.StartsWith(f, StringComparison.OrdinalIgnoreCase) || s.StartsWith(field, StringComparison.OrdinalIgnoreCase));
+    if (tertiaryField != null) return tertiaryField;
+    zdoField = true;
+    var quaternaryField = zdoFields.FirstOrDefault(s => s.StartsWith(field, StringComparison.OrdinalIgnoreCase));
+    if (quaternaryField != null) return quaternaryField;
     return field;
   }
+
+  private static void InitZdoFields()
+  {
+    if (ZdoFields.ContainsKey("zdo")) return;
+    if (WorldEditCommands.IsCLLC)
+    {
+      var values = ZdoFields["MonsterAI"];
+      values["CL&LC effect"] = typeof(Enum_CLLC_Effect);
+      values["CL&LC infusion"] = typeof(Enum_CLLC_Infusion);
+    }
+    Dictionary<string, Type> zdoValues = [];
+    foreach (var kvp in ZdoFields)
+    {
+      foreach (var kvp2 in kvp.Value)
+        zdoValues[kvp2.Key] = kvp2.Value;
+    }
+    ZdoFields["zdo"] = zdoValues;
+  }
+  private static readonly Dictionary<string, Dictionary<string, Type>> ZdoFields = new(){
+    {
+      nameof(ArmorStand), new Dictionary<string, Type>
+      {
+        { "pose", typeof(int) },
+        { "0_item", typeof(GameObject) },
+        { "1_item", typeof(GameObject) },
+        { "2_item", typeof(GameObject) },
+        { "0_variant", typeof(int) },
+        { "1_variant", typeof(int) },
+        { "2_variant", typeof(int) }
+      }
+    },{
+      nameof(BaseAI), new Dictionary<string, Type>
+      {
+        { "aggravated", typeof(bool) },
+        { "alert", typeof(bool) },
+        { "bosscount", typeof(bool) },
+        { "spawntime", typeof(long) },
+        { "ShownAlertMessage", typeof(bool) },
+        { "huntplayer", typeof(bool) },
+        { "lastWorldTime", typeof(long) },
+        { "spawnpoint", typeof(Vector3) },
+        { "patrol", typeof(bool) },
+        { "patrolPoint", typeof(Vector3) },
+      }
+    },{
+      nameof(Bed), new Dictionary<string, Type>
+      {
+        { "owner", typeof(long) },
+        { "ownerName", typeof(string) },
+      }
+    },{
+      nameof(Beehive), new Dictionary<string, Type>
+      {
+        { "lastTime", typeof(long) },
+        { "level", typeof(int) },
+        { "product", typeof(float) },
+      }
+    },{
+      nameof(Catapult), new Dictionary<string, Type>
+      {
+        { "Locked", typeof(bool) },
+        { "visual", typeof(GameObject) },
+      }
+    },{
+      nameof(Character), new Dictionary<string, Type>
+      {
+        { "BodyVelocity", typeof(Vector3) },
+        { "bosscount", typeof(bool) },
+        { "health", typeof(float) },
+        { "level", typeof(int) },
+        { "max_health", typeof(float) },
+        { "noise", typeof(float) },
+        { "RandomSkillFactor", typeof(float) },
+        { "tamed", typeof(bool) },
+      }
+    },{
+      nameof(CharacterAnimEvent), new Dictionary<string, Type>
+      {
+        { "LookTarget", typeof(Vector3) },
+      }
+    },{
+      nameof(Cinder), new Dictionary<string, Type>
+      {
+        { "spread", typeof(bool) },
+      }
+    },{
+      nameof(CinderSpawner), new Dictionary<string, Type>
+      {
+        { "spread", typeof(bool) },
+      }
+    },{
+      nameof(Container), new Dictionary<string, Type>
+      {
+        { "addedDefaultItems", typeof(bool) },
+        { "InUse", typeof(bool) },
+        { "items", typeof(string) },
+      }
+    },{
+      nameof(CookingStation), new Dictionary<string, Type>
+      {
+        { "fuel", typeof(float) },
+        { "StartTime", typeof(long) },
+        { "slot0", typeof(GameObject) },
+        { "slot1", typeof(GameObject) },
+        { "slot2", typeof(GameObject) },
+        { "slotstatus0", typeof(int) },
+        { "slotstatus1", typeof(int) },
+        { "slotstatus2", typeof(int) },
+      }
+    },{
+      nameof(Corpse), new Dictionary<string, Type>
+      {
+        { "timeOfDeath", typeof(long) },
+      }
+    },{
+      nameof(CreatureSpawner), new Dictionary<string, Type>
+      {
+        { "alive_time", typeof(long) },
+      }
+    },{
+      nameof(DungeonGenerator), new Dictionary<string, Type>
+      {
+        { "rooms", typeof(int) },
+        { "room0", typeof(RoomHash) },
+        { "room1", typeof(RoomHash) },
+        { "room2", typeof(RoomHash) },
+        { "room0_pos", typeof(Vector3) },
+        { "room1_pos", typeof(Vector3) },
+        { "room2_pos", typeof(Vector3) },
+        { "room0_rot", typeof(Quaternion) },
+        { "room1_rot", typeof(Quaternion) },
+        { "room2_rot", typeof(Quaternion) },
+      }
+    },{
+      nameof(Destructible), new Dictionary<string, Type>
+      {
+        { "health", typeof(float) },
+      }
+    },{
+      nameof(Door), new Dictionary<string, Type>
+      {
+        { "state", typeof(int) },
+      }
+    },{
+      nameof(EggGrow), new Dictionary<string, Type>
+      {
+        { "GrowStart", typeof(float) },
+      }
+    },{
+      nameof(Fermenter), new Dictionary<string, Type>
+      {
+        { "Content", typeof(GameObject) },
+        { "StartTime", typeof(long) },
+      }
+    },{
+      nameof(Fireplace), new Dictionary<string, Type>
+      {
+        { "fuel", typeof(float) },
+        { "lastTime", typeof(long) },
+      }
+    },{
+      nameof(Fish), new Dictionary<string, Type>
+      {
+        { "escape", typeof(float) },
+        { "hooked", typeof(bool) },
+        { "spawnpoint", typeof(Vector3) },
+      }
+    },{
+      nameof(FishingFloat), new Dictionary<string, Type>
+      {
+        { "Bait", typeof(GameObject) },
+        { "rodOwner", typeof(long) },
+      }
+    },{
+      nameof(Gibber), new Dictionary<string, Type>
+      {
+        { "HitPoint", typeof(Vector3) },
+        { "HitDir", typeof(Vector3) },
+      }
+    },{
+      nameof(Humanoid), new Dictionary<string, Type>
+      {
+        { "IsBlocking", typeof(bool) },
+        { "seed", typeof(int) },
+      }
+    },{
+      nameof(ItemDrop), new Dictionary<string, Type>
+      {
+        { "crafterID", typeof(long) },
+        { "crafterName", typeof(string) },
+        { "data_0", typeof(string) },
+        { "data_1", typeof(string) },
+        { "data_2", typeof(string) },
+        { "data__0", typeof(string) },
+        { "data__1", typeof(string) },
+        { "data__2", typeof(string) },
+        { "durability", typeof(float) },
+        { "pickedUp", typeof(bool) },
+        { "quality", typeof(int) },
+        { "spawntime", typeof(long) },
+        { "stack", typeof(int) },
+        { "variant", typeof(int) },
+        { "worldLevel", typeof(int) },
+      }
+    },{
+      nameof(ItemStand), new Dictionary<string, Type>
+      {
+        { "quality", typeof(int) },
+        { "item", typeof(GameObject) },
+        { "variant", typeof(int) },
+      }
+    },{
+      nameof(Leviathan), new Dictionary<string, Type>
+      {
+        { "dead", typeof(bool) },
+      }
+    },{
+      nameof(LineConnect), new Dictionary<string, Type>
+      {
+        { "line_slack", typeof(float) },
+      }
+    },{
+      nameof(LocationProxy), new Dictionary<string, Type>
+      {
+        { "location", typeof(LocationHash) },
+        { "seed", typeof(int) },
+      }
+    },{
+      nameof(LootSpawner), new Dictionary<string, Type>
+      {
+        { "spawntime", typeof(long) },
+      }
+    },{
+      nameof(MineRock), new Dictionary<string, Type>
+      {
+        { "Health0", typeof(float) },
+        { "Health1", typeof(float) },
+        { "Health2", typeof(float) },
+      }
+    },{
+      nameof(MineRock5), new Dictionary<string, Type>
+      {
+        { "health", typeof(float) },
+      }
+    },{
+      nameof(MonsterAI), new Dictionary<string, Type>
+      {
+        { "DespawnInDay", typeof(bool) },
+        { "EventCreature", typeof(bool) },
+        { "sleeping", typeof(bool) },
+      }
+    },{
+      nameof(MusicLocation), new Dictionary<string, Type>
+      {
+        { "played", typeof(bool) },
+      }
+    },{
+       nameof(Pickable), new Dictionary<string, Type>
+      {
+        { "picked", typeof(bool) },
+        { "picked_time", typeof(long) },
+      }
+    },{
+       nameof(PickableItem), new Dictionary<string, Type>
+      {
+        { "itemPrefab", typeof(ObjectHash) },
+        { "itemStack", typeof(int) },
+      }
+    },{
+      nameof(Piece), new Dictionary<string, Type>
+      {
+        { "creator", typeof(long) },
+      }
+    },{
+      nameof(Plant), new Dictionary<string, Type>
+      {
+        { "seed", typeof(int) },
+        { "plantTime", typeof(long) },
+      }
+    },{
+      nameof(Player), new Dictionary<string, Type>
+      {
+        { "baseValue", typeof(int) },
+        { "DebugFly", typeof(bool) },
+        { "dodgeinv", typeof(bool) },
+        { "dead", typeof(bool) },
+        { "eitr", typeof(float) },
+        { "emote", typeof(string) },
+        { "emoteID", typeof(int) },
+        { "emote_oneshot", typeof(bool) },
+        { "inBed", typeof(bool) },
+        { "playerID", typeof(long) },
+        { "playerName", typeof(string) },
+        { "pvp", typeof(bool) },
+        { "stamina", typeof(float) },
+        { "Stealth", typeof(float) },
+        { "wakeup", typeof(bool) },
+        { "WeaponLoaded", typeof(bool) },
+      }
+    },{
+      nameof(PrivateArea), new Dictionary<string, Type>
+      {
+        { "creatorName", typeof(string) },
+        { "enabled", typeof(bool) },
+        { "permitted", typeof(int) },
+        { "pu_id0", typeof(long) },
+        { "pu_id1", typeof(long) },
+        { "pu_id2", typeof(long) },
+        { "pu_name0", typeof(string) },
+        { "pu_name1", typeof(string) },
+        { "pu_name2", typeof(string) },
+      }
+    },{
+      nameof(Procreation), new Dictionary<string, Type>
+      {
+        { "lovePoints", typeof(int) },
+        { "pregnant", typeof(long) },
+      }
+    },{
+      nameof(Projectile), new Dictionary<string, Type>
+      {
+        { "visual", typeof(string) },
+      }
+    },{
+      nameof(Ragdoll), new Dictionary<string, Type>
+      {
+        { "drops", typeof(int) },
+        { "drop_hash0", typeof(ObjectHash) },
+        { "drop_hash1", typeof(ObjectHash) },
+        { "drop_hash2", typeof(ObjectHash) },
+        { "drop_amount0", typeof(float) },
+        { "drop_amount1", typeof(float) },
+        { "drop_amount2", typeof(float) },
+        { "Hue", typeof(float) },
+        { "InitVel", typeof(Vector3) },
+        { "Saturation", typeof(float) },
+        { "Value", typeof(float) },
+      }
+    },{
+      nameof(RandomFlyingBird), new Dictionary<string, Type>
+      {
+        { "landed", typeof(bool) },
+        { "spawnpoint", typeof(Vector3) },
+      }
+    },{
+      nameof(ResourceRoot), new Dictionary<string, Type>
+      {
+        { "lastTime", typeof(long) },
+        { "level", typeof(float) },
+      }
+    },{
+      nameof(Sadle), new Dictionary<string, Type>
+      {
+        { "user", typeof(long) },
+      }
+    },{
+      nameof(SapCollector), new Dictionary<string, Type>
+      {
+        { "lastTime", typeof(long) },
+        { "level", typeof(int) },
+        { "product", typeof(float) },
+      }
+    },{
+      nameof(SEMan), new Dictionary<string, Type>
+      {
+        { "seAttrib", typeof(int) },
+      }
+    },{
+      nameof(ShieldGenerator), new Dictionary<string, Type>
+      {
+        { "fuel", typeof(float) },
+        { "StartTime", typeof(long) },
+      }
+    },{
+      nameof(Ship), new Dictionary<string, Type>
+      {
+        { "forward", typeof(int) },
+        { "rudder", typeof(float) },
+      }
+    },{
+      nameof(ShipConstructor), new Dictionary<string, Type>
+      {
+        { "user", typeof(long) },
+      }
+    },{
+      nameof(Sign), new Dictionary<string, Type>
+      {
+        { "author", typeof(string) },
+        { "text", typeof(string) },
+      }
+    },{
+      nameof(Smelter), new Dictionary<string, Type>
+      {
+        { "accTime", typeof(float) },
+        { "bakeTimer", typeof(float) },
+        { "fuel", typeof(float) },
+        { "item0", typeof(GameObject) },
+        { "item1", typeof(GameObject) },
+        { "item2", typeof(GameObject) },
+        { "queued", typeof(int) },
+        { "SpawnOre", typeof(GameObject) },
+        { "SpawnAmount", typeof(int) },
+        { "StartTime", typeof(long) },
+      }
+    },{
+      nameof(Tameable), new Dictionary<string, Type>
+      {
+        { "HaveSaddle", typeof(bool) },
+        { "TamedName", typeof(string) },
+        { "TamedNameAuthor", typeof(string) },
+      }
+    },{
+      nameof(TeleportWorld), new Dictionary<string, Type>
+      {
+        { "tag", typeof(string) },
+        { "tagauthor", typeof(string) },
+      }
+    },{
+      nameof(Trap), new Dictionary<string, Type>
+      {
+        { "state", typeof(int) },
+        { "triggered", typeof(float) },
+      }
+    },{
+      nameof(TreeBase), new Dictionary<string, Type>
+      {
+        { "health", typeof(float) },
+      }
+    },{
+      nameof(TreeLog), new Dictionary<string, Type>
+      {
+        { "health", typeof(float) },
+      }
+    },{
+      nameof(TriggerSpawner), new Dictionary<string, Type>
+      {
+        { "spawntime", typeof(long) },
+      }
+    },{
+      nameof(Turret), new Dictionary<string, Type>
+      {
+        { "ammo", typeof(int) },
+        { "ammoType", typeof(GameObject) },
+        { "lastAttack", typeof(float) },
+        { "targets", typeof(int) },
+        { "target0", typeof(string) },
+        { "target1", typeof(string) },
+        { "target2", typeof(string) },
+      }
+    },{
+      nameof(Vagon), new Dictionary<string, Type>
+      {
+        { "attachJoint", typeof(bool) },
+      }
+    },{
+      nameof(WearNTear), new Dictionary<string, Type>
+      {
+        { "health", typeof(float) },
+        { "support", typeof(float) },
+      }
+    },{
+      nameof(WispSpawner), new Dictionary<string, Type>
+      {
+        { "LastSpawn", typeof(long) },
+      }
+    },
+  };
 }
 
 [HarmonyPatch(typeof(ZDO))]