+++ WorldEditCommands/tweak/TweakCreature.cs
@@ -10,28 +10,30 @@ public class TweakCreatureCommand : TweakCommand
   {
     if (operation == "name") return TweakActions.Name(view, value);
     if (operation == "faction") return TweakActions.Faction(view, value);
-    throw new System.NotImplementedException();
+    throw new NotImplementedException();
   }
   protected override string DoOperation(ZNetView view, string operation, float? value)
   {
     if (operation == "health") return TweakActions.Health(view, value);
     if (operation == "damage") return Actions.Damage(view, value);
-    throw new System.NotImplementedException();
+    throw new NotImplementedException();
   }
 
   protected override string DoOperation(ZNetView view, string operation, int? value)
   {
     if (operation == "level") return TweakActions.Level(view, value);
-    throw new System.NotImplementedException();
+    throw new NotImplementedException();
   }
 
   protected override string DoOperation(ZNetView view, string operation, string[] value)
   {
     if (operation == "item") return TweakActions.Items(view, value);
-    if (operation == "affix") return TweakActions.CLLC(view, value);
+    if (operation == "affix") return TweakActions.CLLC_BossEffect(view, value);
+    if (operation == "effect") return TweakActions.CLLC_Effect(view, value);
+    if (operation == "infusion") return TweakActions.CLLC_Infusion(view, value);
     if (operation == "resistance") return TweakActions.Resistances(view, value);
     if (operation == "attacks") return TweakActions.Attacks(view, value);
-    throw new System.NotImplementedException();
+    throw new NotImplementedException();
   }
 
   protected override string DoOperation(ZNetView view, string operation, bool? value)
@@ -66,12 +68,20 @@ public class TweakCreatureCommand : TweakCommand
       SupportedOperations.Add("attacks", typeof(string[]));
     }
     if (WorldEditCommands.IsCLLC)
+    {
       SupportedOperations.Add("affix", typeof(string[]));
+      SupportedOperations.Add("effect", typeof(string[]));
+      SupportedOperations.Add("infusion", typeof(string[]));
+    }
 
     if (WorldEditCommands.IsCLLC)
-      AutoComplete.Add("affix", (int index) => Enum.GetNames(typeof(BossAffix)).ToList());
+    {
+      AutoComplete.Add("affix", (int index) => [.. Enum.GetNames(typeof(Enum_CLLC_Boss))]);
+      AutoComplete.Add("effect", (int index) => [.. Enum.GetNames(typeof(Enum_CLLC_Effect))]);
+      AutoComplete.Add("infusion", (int index) => [.. Enum.GetNames(typeof(Enum_CLLC_Infusion))]);
+    }
     AutoComplete.Add("name", (int index) => index == 0 ? ParameterInfo.Create("name=<color=yellow>text</color>", "Display name. Use _ as the space. No value to reset.") : ParameterInfo.None);
-    AutoComplete.Add("faction", (int index) => index == 0 ? Enum.GetNames(typeof(Character.Faction)).ToList() : ParameterInfo.None);
+    AutoComplete.Add("faction", (int index) => index == 0 ? [.. Enum.GetNames(typeof(Character.Faction))] : ParameterInfo.None);
     AutoComplete.Add("boss", (int index) => ParameterInfo.Create("boss=<color=yellow>true/false</color> ", "Sets the boss health bar. No value to reset."));
     AutoComplete.Add("tame", (int index) => ParameterInfo.Create("tame=<color=yellow>true/false</color> ", "Sets the tamed status. No value to toggle."));
     AutoComplete.Add("hunt", (int index) => ParameterInfo.Create("hunt=<color=yellow>true/false</color>", "Sets the extra aggressiveness. No value to toggle."));
@@ -80,8 +90,8 @@ public class TweakCreatureCommand : TweakCommand
     AutoComplete.Add("damage", (int index) => ParameterInfo.Create("damage=<color=yellow>number</color>", "Sets the damage multiplier."));
     AutoComplete.Add("resistance", (int index) =>
     {
-      if (index == 0) return Enum.GetNames(typeof(HitData.DamageType)).ToList();
-      if (index == 1) return Enum.GetNames(typeof(HitData.DamageModifier)).ToList();
+      if (index == 0) return [.. Enum.GetNames(typeof(HitData.DamageType))];
+      if (index == 1) return [.. Enum.GetNames(typeof(HitData.DamageModifier))];
       return ParameterInfo.Create("For additional entries, add more <color>resistance=...</color> parameters.");
     });
     AutoComplete.Add("item", (int index) =>